pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autopr-engine"
version = "1.0.0"
description = "AI-Powered GitHub PR Automation and Issue Management"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "VeritasVault Team", email = "dev@veritasvault.net"}
]
maintainers = [
    {name = "VeritasVault Team", email = "dev@veritasvault.net"}
]
keywords = [
    "github", "pull-request", "automation", "ai", "code-review",
    "ci-cd", "workflow", "integration", "slack", "linear",
    "autogen", "llm", "openai", "anthropic", "issue-management",
    "quality-gates", "platform-detection", "multi-agent"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    # Core dependencies
    "pydantic>=2.0.0,<3.0.0",
    "aiohttp>=3.8.0,<4.0.0",
    "structlog>=22.0.0,<24.0.0",
    "click>=8.0.0,<9.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "jinja2>=3.1.0,<4.0.0",
    
    # GitHub integration
    "pygithub>=1.58.0,<2.0.0",
    "GitPython>=3.1.0,<4.0.0",
    
    # AI and LLM providers
    "openai>=1.0.0,<2.0.0",
    "anthropic>=0.25.0,<1.0.0",
    "mistralai>=0.4.0,<1.0.0",
    
    # HTTP and networking
    "httpx>=0.24.0,<1.0.0",
    "websockets>=11.0.0,<12.0.0",
    
    # Data processing
    "python-dateutil>=2.8.0,<3.0.0",
    "pytz>=2023.3",
    
    # Configuration and environment
    "python-dotenv>=1.0.0,<2.0.0",
    "toml>=0.10.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-mock>=3.11.0,<4.0.0",
    "black>=23.7.0,<24.0.0",
    "isort>=5.12.0,<6.0.0",
    "flake8>=6.0.0,<7.0.0",
    "mypy>=1.5.0,<2.0.0",
    "pre-commit>=3.4.0,<4.0.0",
    "sphinx>=7.1.0,<8.0.0",
    "sphinx-rtd-theme>=1.3.0,<2.0.0",
]
monitoring = [
    "prometheus_client>=0.17.0,<1.0.0",
    "sentry-sdk[fastapi]>=1.32.0,<2.0.0",
    "datadog>=0.47.0,<1.0.0",
]
memory = [
    "mem0ai>=0.1.0,<1.0.0",
    "chromadb>=0.4.0,<1.0.0",
    "qdrant-client>=1.5.0,<2.0.0",
]
ai = [
    "pyautogen>=0.2.0,<1.0.0",
    "langchain>=0.0.300,<1.0.0",
    "langchain-openai>=0.0.5,<1.0.0",
    "langchain-anthropic>=0.1.0,<1.0.0",
]
database = [
    "asyncpg>=0.28.0,<1.0.0",
    "sqlalchemy[asyncio]>=2.0.0,<3.0.0",
    "alembic>=1.12.0,<2.0.0",
    "redis>=4.6.0,<5.0.0",
    "aioredis>=2.0.0,<3.0.0",
]
server = [
    "fastapi>=0.103.0,<1.0.0",
    "uvicorn[standard]>=0.23.0,<1.0.0",
    "gunicorn>=21.2.0,<22.0.0",
]
resilience = [
    "pybreaker>=1.0.0,<2.0.0",
    "tenacity>=8.2.0,<9.0.0",
    "limits>=3.6.0,<4.0.0",
]
full = [
    "autopr-engine[dev,monitoring,memory,ai,database,server,resilience]"
]

[project.urls]
Homepage = "https://github.com/veritasvault/autopr-engine"
Documentation = "https://autopr-engine.readthedocs.io"
Repository = "https://github.com/veritasvault/autopr-engine"
"Bug Tracker" = "https://github.com/veritasvault/autopr-engine/issues"
Changelog = "https://github.com/veritasvault/autopr-engine/blob/main/CHANGELOG.md"
Discussions = "https://github.com/veritasvault/autopr-engine/discussions"

[project.scripts]
autopr = "autopr.cli:main"
autopr-server = "autopr.server:main"
autopr-worker = "autopr.worker:main"
autopr-migration = "autopr.migration:main"

[project.entry-points."autopr.actions"]
platform_detector = "autopr.actions.platform_detector:PlatformDetector"
pr_review_analyzer = "autopr.actions.pr_review_analyzer:PRReviewAnalyzer"
issue_creator = "autopr.actions.issue_creator:IssueCreator"
ai_comment_analyzer = "autopr.actions.ai_comment_analyzer:AICommentAnalyzer"
quality_gates = "autopr.actions.quality_gates:QualityGates"
autogen_multi_agent = "autopr.actions.autogen_multi_agent:AutoGenMultiAgent"

[project.entry-points."autopr.integrations"]
github = "autopr.integrations.github:GitHubIntegration"
linear = "autopr.integrations.linear:LinearIntegration"
slack = "autopr.integrations.slack:SlackIntegration"
axolo = "autopr.integrations.axolo:AxoloIntegration"

[project.entry-points."autopr.llm_providers"]
openai = "autopr.ai.providers.openai:OpenAIProvider"
anthropic = "autopr.ai.providers.anthropic:AnthropicProvider"
mistral = "autopr.ai.providers.mistral:MistralProvider"
groq = "autopr.ai.providers.groq:GroqProvider"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
autopr = [
    "workflows/*.yaml",
    "workflows/*.yml", 
    "config/*.yaml",
    "config/*.yml",
    "templates/*.j2",
    "templates/*.jinja",
    "static/*",
]

# Development tools configuration
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["autopr"]
known_third_party = [
    "pydantic", "aiohttp", "structlog", "click", "yaml",
    "openai", "anthropic", "autogen", "github", "linear"
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "autogen.*",
    "github.*",
    "linear.*",
    "mem0.*",
    "qdrant_client.*",
    "chromadb.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=autopr",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["autopr"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
per-file-ignores = [
    "__init__.py:F401",
]

# Documentation configuration
[tool.sphinx]
project = "AutoPR Engine"
copyright = "2025, VeritasVault Team"
author = "VeritasVault Team"
extensions = [
    "sphinx.ext.autodoc",
    "sphinx.ext.viewcode",
    "sphinx.ext.napoleon",
    "sphinx.ext.intersphinx",
    "sphinx.ext.autosummary",
]
templates_path = ["_templates"]
exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
html_theme = "sphinx_rtd_theme"
html_static_path = ["_static"] 